+++ quantify_scheduler/operations/measurement_factories.py
@@ -6,7 +6,7 @@ A module containing factory functions for measurements on the quantum-device lay
 These factories are used to take a parametrized representation of on a operation
 and use that to create an instance of the operation itself.
 """
-from typing import Literal, Union
+from typing import List, Literal, Union
 
 from quantify_scheduler import Operation
 from quantify_scheduler.enums import BinMode
@@ -109,10 +109,10 @@ def dispersive_measurement(
 
 
 def optical_measurement(
-    pulse_amplitude: float,
-    pulse_duration: float,
-    pulse_port: str,
-    pulse_clock: str,
+    pulse_amplitudes: List[float],
+    pulse_durations: List[float],
+    pulse_ports: List[str],
+    pulse_clocks: List[str],
     acq_duration: float,
     acq_delay: float,
     acq_port: str,
@@ -120,28 +120,33 @@ def optical_measurement(
     acq_channel: int,
     acq_index: int,
     bin_mode: Union[BinMode, None],
-    acq_protocol: Literal["Trace", "TriggerCount"],
+    acq_protocol: Literal["Trace", "TriggerCount", None],
     acq_protocol_default: Literal["Trace", "TriggerCount"],
     pulse_type: Literal["SquarePulse"],
 ) -> Operation:
     # pylint: disable=too-many-locals
-    """Generator function for a standard optical measurement.
+    """Generator function for an optical measurement with multiple excitation pulses.
 
-    An optical measurement generates a square pulse in the optical range and uses the
-    Trace acquisition to return the output of a photon detector as a function of time.
-    Alternatively, the TriggerCount counts the number of photons that are collected.
+    An optical measurement generates a square pulse in the optical range and uses
+    either the Trace acquisition to return the output of a photon detector as a
+    function of time or the TriggerCount acquisition to return the number of photons
+    that are collected.
 
+    All pulses can have different amplitudes, durations, ports and clocks. All pulses
+    start simultaneously. The acquisition can have an ``acq_delay`` with respect to the
+    pulses. A negative ``acq_delay`` causes the acquisition to be scheduled at time 0
+    and the pulses at the positive time ``-acq_delay``.
 
     Parameters
     ----------
-    pulse_amplitude
-        Amplitude of the generated pulse
-    pulse_duration
-        Duration of the generated pulse
-    pulse_port
-        Port name, where the pulse is applied
-    pulse_clock
-        Clock name of the generated pulse
+    pulse_amplitudes
+        list of amplitudes of the corresponding pulses
+    pulse_durations
+        list of durations of the corresponding pulses
+    pulse_ports
+        Port names, where the corresponding pulses are applied
+    pulse_clocks
+        Clock names of the corresponding pulses
     acq_duration
         Duration of the acquisition
     acq_delay
@@ -164,46 +169,68 @@ def optical_measurement(
     acq_protocol
         Acquisition protocol. "Trace" returns a time trace of the collected signal.
         "TriggerCount" returns the number of times the trigger threshold is surpassed.
-    acq_protocol_default, optional
-        Acquisition protocol if ``acq_protocol`` is None, by default "TriggerCount"
-    pulse_type, optional
-        Shape of the pulse to be generated, by default "SquarePulse"
+    acq_protocol_default
+        Acquisition protocol if ``acq_protocol`` is None
+    pulse_type
+        Shape of the pulse to be generated
 
     Returns
     -------
-        Operation with the generated pulse and acquisition
+        Operation with the generated pulses and acquisition
 
     Raises
     ------
+    ValueError
+        If first four function arguments do not have the same length.
     NotImplementedError
         If an unknown ``pulse_type`` or ``acq_protocol`` are used.
     """
+
     # ensures default argument is used if not specified at gate level.
     # ideally, this input would not be accepted, but this is a workaround for #267
     if bin_mode is None:
         bin_mode = BinMode.APPEND
 
+    # All lists should be of equal length so this should be ensured
+    if (
+        not len(pulse_amplitudes)
+        == len(pulse_durations)
+        == len(pulse_ports)
+        == len(pulse_clocks)
+    ):
+        raise ValueError(
+            "For multiple optical excitations, lists must have same length:\n"
+            + f"{len(pulse_amplitudes)=},\n"
+            + f"{len(pulse_durations)=},\n"
+            + f"{len(pulse_ports)=},\n"
+            + f"{len(pulse_clocks)=}"
+        )
+
     # If acq_delay >= 0, the pulse starts at 0 and the acquisition at acq_delay
-    # If acq_delay < 0, the pulse starts at -acq_delay and the acquisition at 0
+    # If acq_delay < 0, the acquisition starts at 0 and the pulse at -acq_delay (which is positive)
     t0_pulse = max(0, -acq_delay)
     t0_acquisition = max(0, acq_delay)
 
+    # This operation will contain all pulses and the acquisition
     device_op = Operation("OpticalMeasurement")
+
     if pulse_type == "SquarePulse":
-        device_op.add_pulse(
-            SquarePulse(
-                amp=pulse_amplitude,
-                duration=pulse_duration,
-                port=pulse_port,
-                clock=pulse_clock,
-                t0=t0_pulse,
+        settings = zip(pulse_amplitudes, pulse_durations, pulse_ports, pulse_clocks)
+        for amp, dur, port, clock in settings:
+            device_op.add_pulse(
+                SquarePulse(
+                    amp=amp,
+                    duration=dur,
+                    port=port,
+                    clock=clock,
+                    t0=t0_pulse,
+                )
             )
-        )
     else:
         raise NotImplementedError(
-            f'Invalid pulse_type "{pulse_type}" specified as argument to '
-            'optical_measurement. Currently, only "SquarePulse" is accepted. '
-            "Please correct your device config."
+            f"Invalid pulse_type '{pulse_type}' specified as argument to "
+            f"optical_measurement. Currently, only 'SquarePulse' is accepted. "
+            f"Please correct your device config."
         )
 
     if acq_protocol is None:
@@ -234,8 +261,8 @@ def optical_measurement(
         )
     else:
         raise NotImplementedError(
-            f'Acquisition protocol "{acq_protocol}" is not supported. '
-            'Currently, only "TriggerCount" and "Trace" are accepted.'
+            f"Acquisition protocol '{acq_protocol}' is not supported. "
+            f"Currently, only 'TriggerCount' and 'Trace' are accepted."
         )
 
     return device_op