+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - NV centers - `Reset` operation with compilation for Qblox hardware (!485)
 - NV centers - `Measure` operation using TriggerCount acquisition; only device compilation so far (!490)
 - NV Centers - `ChargeReset` operation only device compilation so far (!496)
+- NV centers - `CRCount` operation using TriggerCount acquisition; only device compilation so far (!502)
 - QBlox backend, Operations - The phase of the measurement clock is set to zero at the start of each measurement by default (!434, #296)
 - Operations - ShiftClockPhase now uses `phase_shift` as keyword instead of `phase` (!434)
 - Qblox backend - Introduce `"sequence_to_file"` param in qblox hardware config to allow skipping writing sequence json files to disk (#108, !438)